+++ packages/cubejs-postgres-driver/driver/index.d.ts
@@ -1,0 +1,8 @@
+import { PoolConfig } from "pg";
+
+declare module "@cubejs-backend/postgres-driver" {
+  class PostgresDriver {
+    constructor(options?: PoolConfig);
+  }
+  export = PostgresDriver;
+}