+++ packages/cubejs-schema-compiler/adapter/PreAggregations.js
@@ -89,17 +89,34 @@ class PreAggregations {
       loadSql: this.query.preAggregationLoadSql(cube, preAggregation, tableName),
       invalidateKeyQueries: refreshKeyQueries.queries,
       refreshKeyRenewalThresholds: refreshKeyQueries.refreshKeyRenewalThresholds,
-      external: preAggregation.external
+      external: preAggregation.external,
+      indexesSql: Object.keys(preAggregation.indexes || {}).map(
+        index => {
+          const indexName = this.preAggregationTableName(cube, `${preAggregationName}_${index}`, preAggregation, true);
+          return {
+            indexName,
+            sql: this.query.indexSql(
+              cube,
+              preAggregation,
+              preAggregation.indexes[index],
+              indexName,
+              tableName
+            )
+          };
+        }
+      )
     };
   }
 
-  preAggregationTableName(cube, preAggregationName, preAggregation) {
+  preAggregationTableName(cube, preAggregationName, preAggregation, skipSchema) {
     return this.query.preAggregationTableName(
       cube, preAggregationName + (
-      preAggregation.partitionTimeDimensions ?
-        preAggregation.partitionTimeDimensions[0].dateRange[0].replace('T00:00:00.000', '').replace(/-/g, '') :
-        ''
-    ));
+        preAggregation.partitionTimeDimensions ?
+          preAggregation.partitionTimeDimensions[0].dateRange[0].replace('T00:00:00.000', '').replace(/-/g, '') :
+          ''
+      ),
+      skipSchema
+    );
   }
 
   findPreAggregationToUseForCube(cube) {
@@ -148,6 +165,7 @@ class PreAggregations {
     const isAdditive = R.all(m => m.isAdditive(), query.measures);
     const leafMeasureAdditive = R.all(path => query.newMeasure(path).isAdditive(), leafMeasurePaths);
     const granularityHierarchies = query.granularityHierarchies();
+    const hasMultipliedMeasures = query.fullKeyQueryAggregateMeasures().multipliedMeasures.length > 0;
 
     return {
       sortedDimensions,
@@ -158,7 +176,8 @@ class PreAggregations {
       hasNoTimeDimensionsWithoutGranularity,
       allFiltersWithinSelectedDimensions,
       isAdditive,
-      granularityHierarchies
+      granularityHierarchies,
+      hasMultipliedMeasures
     };
   }
 
@@ -215,8 +234,7 @@ class PreAggregations {
     // [[TimeDimension]]
     const queryTimeDimensionsList = transformedQuery.sortedTimeDimensions.map(expandTimeDimension);
 
-    const canUsePreAggregationNotAdditive = (references) =>
-      transformedQuery.hasNoTimeDimensionsWithoutGranularity &&
+    const canUsePreAggregationNotAdditive = (references) => transformedQuery.hasNoTimeDimensionsWithoutGranularity &&
       transformedQuery.allFiltersWithinSelectedDimensions &&
       R.equals(references.sortedDimensions || references.dimensions, transformedQuery.sortedDimensions) &&
       (
@@ -228,35 +246,18 @@ class PreAggregations {
         references.sortedTimeDimensions || sortTimeDimensions(references.timeDimensions)
       );
 
-    const canUsePreAggregationLeafMeasureAdditive = (references) =>
-      R.all(
-        d => (references.sortedDimensions || references.dimensions).indexOf(d) !== -1,
-        transformedQuery.sortedDimensions
-      ) &&
+    const canUsePreAggregationLeafMeasureAdditive = (references) => R.all(
+      d => (references.sortedDimensions || references.dimensions).indexOf(d) !== -1,
+      transformedQuery.sortedDimensions
+    ) &&
       R.all(m => references.measures.indexOf(m) !== -1, transformedQuery.leafMeasures) &&
       R.allPass(
         queryTimeDimensionsList.map(tds => R.anyPass(tds.map(td => R.contains(td))))
       )(references.sortedTimeDimensions || sortTimeDimensions(references.timeDimensions));
 
-    const canUsePreAggregationAdditive = (references) =>
-      R.all(
-        d => (references.sortedDimensions || references.dimensions).indexOf(d) !== -1,
-        transformedQuery.sortedDimensions
-      ) &&
-      (
-        R.all(m => references.measures.indexOf(m) !== -1, transformedQuery.measures) ||
-        R.all(m => references.measures.indexOf(m) !== -1, transformedQuery.leafMeasures)
-      ) &&
-      R.allPass(
-        queryTimeDimensionsList.map(tds => R.anyPass(tds.map(td => R.contains(td))))
-      )(references.sortedTimeDimensions || sortTimeDimensions(references.timeDimensions));
-
-
     let canUseFn;
-    if (transformedQuery.isAdditive) {
-      canUseFn = canUsePreAggregationAdditive;
-    } else if (transformedQuery.leafMeasureAdditive) {
-      canUseFn = canUsePreAggregationLeafMeasureAdditive;
+    if (transformedQuery.leafMeasureAdditive && !transformedQuery.hasMultipliedMeasures) {
+      canUseFn = (r) => canUsePreAggregationLeafMeasureAdditive(r) || canUsePreAggregationNotAdditive(r);
     } else {
       canUseFn = canUsePreAggregationNotAdditive;
     }