+++ packages/cubejs-server-core/core/agentCollect.js
@@ -1,21 +1,14 @@
-import { fetch } from 'whatwg-fetch';
-import cookie from 'component-cookie';
-import uuidv4 from 'uuid/v4';
+const fetch = require('node-fetch');
+const crypto = require('crypto');
 
 let flushPromise = null;
 let trackEvents = [];
-let baseProps = {};
 
-const track = async (event) => {
-  if (!cookie('playground_anonymous')) {
-    cookie('playground_anonymous', uuidv4());
-  }
+module.exports = async (event, endpointUrl, logger) => {
   trackEvents.push({
-    ...baseProps,
     ...event,
-    id: uuidv4(),
-    clientAnonymousId: cookie('playground_anonymous'),
-    clientTimestamp: new Date().toJSON()
+    id: crypto.randomBytes(16).toString('hex'),
+    timestamp: new Date().toJSON()
   });
   const flush = async (toFlush, retries) => {
     if (!toFlush) {
@@ -30,7 +23,7 @@ const track = async (event) => {
     }
     try {
       const sentAt = new Date().toJSON();
-      const result = await fetch('https://track.cube.dev/track', {
+      const result = await fetch(endpointUrl, {
         method: 'post',
         body: JSON.stringify(toFlush.map(r => ({ ...r, sentAt }))),
         headers: { 'Content-Type': 'application/json' },
@@ -43,7 +36,7 @@ const track = async (event) => {
       if (retries > 0) {
         return flush(toFlush, retries - 1);
       }
-      // console.log(e);
+      logger('Agent Error', { error: (e.stack || e).toString() });
     }
     return null;
   };
@@ -54,21 +47,4 @@ const track = async (event) => {
   });
   flushPromise = currentPromise;
   return flushPromise;
-};
-
-export const setAnonymousId = (anonymousId, props) => {
-  baseProps = props;
-  track({ event: 'identify', anonymousId, ...props });
-};
-
-export const event = (name, params) => {
-  track({ event: name, ...params });
-};
-
-export const playgroundAction = (name, options) => {
-  event('Playground Action', { name, ...options });
-};
-
-export const page = (path) => {
-  track({ event: 'page', ...path });
 };